Refactor backfill modal logic into separate logic
