+++ frontend/src/scenes/pipeline/BatchExportRuns.tsx
@@ -4,7 +4,6 @@ import clsx from 'clsx'
 import { useActions, useValues } from 'kea'
 import { DateFilter } from 'lib/components/DateFilter/DateFilter'
 import { NotFound } from 'lib/components/NotFound'
-import { PageHeader } from 'lib/components/PageHeader'
 import { TZLabel } from 'lib/components/TZLabel'
 import { IconCancel, IconRefresh } from 'lib/lemon-ui/icons'
 
@@ -19,7 +18,7 @@ export function BatchExportRuns({ id }: BatchExportRunsLogicProps): JSX.Element 
     const logic = batchExportRunsLogic({ id })
 
     const { batchExportConfig, groupedRuns, loading, hasMoreRunsToLoad, usingLatestRuns } = useValues(logic)
-    const { loadOlderRuns, retryRun, openBackfillModal } = useActions(logic)
+    const { loadOlderRuns, retryRun } = useActions(logic)
 
     if (!batchExportConfig) {
         return <NotFound object="batch export" />
@@ -27,13 +26,6 @@ export function BatchExportRuns({ id }: BatchExportRunsLogicProps): JSX.Element 
 
     return (
         <>
-            <PageHeader
-                buttons={
-                    <LemonButton type="primary" onClick={() => openBackfillModal()}>
-                        Backfill batch export
-                    </LemonButton>
-                }
-            />
             <div className="space-y-2">
                 <BatchExportRunsFilters id={id} />
                 {usingLatestRuns ? (
